Migate TitleCallback to own Trait
