+++ src/Views/Columns/BooleanColumn.php
@@ -6,7 +6,7 @@ use Illuminate\Database\Eloquent\Model;
 use Rappasoft\LaravelLivewireTables\Exceptions\DataTableConfigurationException;
 use Rappasoft\LaravelLivewireTables\Views\Column;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Configuration\BooleanColumnConfiguration;
-use Rappasoft\LaravelLivewireTables\Views\Traits\HasCallback;
+use Rappasoft\LaravelLivewireTables\Views\Traits\Core\HasCallback;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Helpers\BooleanColumnHelpers;
 
 class BooleanColumn extends Column
