+++ src/Views/Columns/LinkColumn.php
@@ -6,18 +6,17 @@ use Illuminate\Database\Eloquent\Model;
 use Rappasoft\LaravelLivewireTables\Exceptions\DataTableConfigurationException;
 use Rappasoft\LaravelLivewireTables\Views\Column;
 use Rappasoft\LaravelLivewireTables\Views\Traits\Configuration\LinkColumnConfiguration;
-use Rappasoft\LaravelLivewireTables\Views\Traits\HasLocationCallback;
+use Rappasoft\LaravelLivewireTables\Views\Traits\Core\{HasLocationCallback,HasTitleCallback};
 use Rappasoft\LaravelLivewireTables\Views\Traits\Helpers\LinkColumnHelpers;
 
 class LinkColumn extends Column
 {
     use LinkColumnConfiguration,
         LinkColumnHelpers,
-        HasLocationCallback;
+        HasLocationCallback,
+        HasTitleCallback;
 
     protected string $view = 'livewire-tables::includes.columns.link';
-
-    protected mixed $titleCallback = null;
 
     public function __construct(string $title, ?string $from = null)
     {
