+++ src/Views/Traits/Configuration/LinkColumnConfiguration.php
@@ -4,11 +4,5 @@ namespace Rappasoft\LaravelLivewireTables\Views\Traits\Configuration;
 
 trait LinkColumnConfiguration
 {
-    // TODO: Test
-    public function title(callable $callback): self
-    {
-        $this->titleCallback = $callback;
 
-        return $this;
-    }
 }
