+++ src/Views/Traits/Helpers/LinkColumnHelpers.php
@@ -4,14 +4,5 @@ namespace Rappasoft\LaravelLivewireTables\Views\Traits\Helpers;
 
 trait LinkColumnHelpers
 {
-    // TODO: Test
-    public function getTitleCallback(): ?callable
-    {
-        return $this->titleCallback;
-    }
 
-    public function hasTitleCallback(): bool
-    {
-        return $this->titleCallback !== null;
-    }
 }
